Moved order-book tests and utils to new directory
